refactored signup, added imports to login
