+++ client/src/index.css
@@ -4,9 +4,7 @@ body {
   font-family: sans-serif;
 }
 
-.nameField,
-.emailField,
-.passwordField {
+.inputField {
   width: 90vw;
 }
